examples/simple/requirements.txt
erichschroeter.code_generator==0.1.1
Jinja2==3.1.4
MarkupSafe==2.1.5
